Use a free list of mrml nodes for createMarkupLine
We maintain a free list, and try to allocate from it first and only
if there are no more free mrmlnodes in the free list do we allocate
more. They are freed back to the free list so that we can reuse them
if we need them again without allocating from Slicer.