Added useFreeList to choose whether to use free list or not, set to False
For now, I am disabling the free list as there is some subtle issue
with reusing a node that is causing annotations to break that I
haven't got to the bottom of. Leaving the free list code in for now
so I can keep debugging while freeing up memory correctly when not
using free list.