Fix leaking overlayVolume and depthGuideVolume
We were just setting these to None without calling RemoveNode()
on them, causing the scene to keep a reference. Then we would
allocate a whole new volume for each in updateOverlayVolume
for depthGuideVolume and in loadNextSequence for overlayVolume,
so the memory would just keep leaking as we drew more lines,
went back and forth in frames etc.

Also made the line endpoints smaller to match Adjudicate, they
are too large currently for small screens.