Use add and remove observers unconditionally
Basically Slicer is losing track or removing observers somewhere
and so just overrride addObserver and removeObserver to allow
adding and removing whether or not the observer is already added
or removed without throwing warnings (it already ignores, but the
warnings make things look bad). Change calls to add and remove
observers to just use a try/except block as all our nodes need
the same observers and so adding them unconditionally is ok when
we create and removing them in freeMarkupNode is also correct.