Add/Remove PointRemoved handler from logic only
It turns out that if you addObserver(..., self.logic.onPointRemoved) from
Widget, you can't removeObserver(..., self.onPointRemoved) from Logic as
the observers don't look the same even though they are calling the same
actual handler. So instead, add the onPointRemoved handler in
createMarkupLine by passing is whether to add it or not as an
argument which is only called from onPasteLines. Normally we only
add the handler after we have both points when adding lines.