Use a Custom Observer Mixin to avoid the warnings on add/remove observers
For some reason, still running into issues trying to use hasObserver before
adding or removing observers. We seem to miss removing some observers when
deleting all nodes and get errors. So now the mixing just allows us to
unconditionally add and remove observers (they are copies of the
VTKObservationMixin) but without the warnings. The warnings were just
warnings, they were not actually re-adding or re-removing anything and
could be ignored, but look ugly and fill up the python console.